feature 9486 finished: Add a trash to resume delete operations on dq elements.modify these file
git-svn-id: http://talendforge.org/svn/top/trunk@45651 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.dataprofiler.core/src/org/talend/dataprofiler/core/ui/action/actions/DeleteFolderAction.java
@@ -22,8 +22,10 @@ import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.ui.ISharedImages;
 import org.eclipse.ui.PlatformUI;
+import org.talend.dataprofiler.core.CorePlugin;
 import org.talend.dataprofiler.core.i18n.internal.DefaultMessagesImpl;
 import org.talend.top.repository.ProxyRepositoryManager;
+import org.talend.dataprofiler.core.recycle.LogicalDeleteFileHandle;
 
 /**
  * DOC qzhang class global comment. Detailled comment <br/>
@@ -36,14 +38,21 @@ public class DeleteFolderAction extends Action {
     protected static Logger log = Logger.getLogger(DeleteFolderAction.class);
 
     private IFolder folder;
+    
+	private boolean isDeleteForever = false;
 
     /**
      * DOC qzhang DeleteFolderAction constructor comment.
      */
-    public DeleteFolderAction(IFolder obj) {
+    public DeleteFolderAction(IFolder obj,boolean isDeleteForever) {
         this.folder = obj;
-        setText(DefaultMessagesImpl.getString("DeleteFolderAction.deleteFolder")); //$NON-NLS-1$
-        setImageDescriptor(PlatformUI.getWorkbench().getSharedImages().getImageDescriptor(ISharedImages.IMG_TOOL_DELETE));
+        this.isDeleteForever = isDeleteForever;
+        if(isDeleteForever){
+        	setText(DefaultMessagesImpl.getString("DeleteObjectsAction.deleteForever")); //$NON-NLS-1$
+        }else{
+        	setText(DefaultMessagesImpl.getString("DeleteFolderAction.deleteFolder")); //$NON-NLS-1$
+        }
+        setImageDescriptor(PlatformUI.getWorkbench().getSharedImages().getImageDescriptor(ISharedImages.IMG_TOOL_DELETE));		
     }
 
     /*
@@ -57,21 +66,43 @@ public class DeleteFolderAction extends Action {
         boolean isFilesDeleted = deleteFolderAndFiles();
         if (!isFilesDeleted) {
             return;
-        }
-        try {
-            folder.delete(true, null);
-            ProxyRepositoryManager.getInstance().save();
-            parent.refreshLocal(IResource.DEPTH_INFINITE, null);
-        } catch (CoreException e) {
-            log.error(e, e);
-        }
+		}
+		try {
+			// MOD qiongli feature 9486
+			if (isDeleteForever) {	
+				delsubFolderForever(folder);
+				if (folder.members().length==0){
+					LogicalDeleteFileHandle.replaceInFile(
+							LogicalDeleteFileHandle.folderType
+									+ folder.getFullPath().toOSString(), "");
+					folder.delete(true, null);
+				}
+			} else {
+				for (IResource member:folder.members()){
+					if (member instanceof IFolder) {
+						LogicalDeleteFileHandle.saveElement(
+								LogicalDeleteFileHandle.folderType,
+								((IFolder) member).getFullPath().toOSString());
+					}
+				}
+				LogicalDeleteFileHandle.saveElement(
+						LogicalDeleteFileHandle.folderType, folder
+								.getFullPath().toOSString());
+			}
+			CorePlugin.getDefault().refreshDQView();
+			// ~
+			ProxyRepositoryManager.getInstance().save();
+			parent.refreshLocal(IResource.DEPTH_INFINITE, null);
+		} catch (CoreException e) {
+			log.error(e, e);
+		}
     }
 
     private boolean deleteFolderAndFiles() {
 
         try {
             if (folder.members().length != 0) {
-                DeleteObjectsAction action = new DeleteObjectsAction();
+                DeleteObjectsAction action = new DeleteObjectsAction(isDeleteForever);
                 action.run();
                 return action.getRunStatus();
             }
@@ -83,6 +114,23 @@ public class DeleteFolderAction extends Action {
                 .getString("DeleteFolderAction.deleteFold"), //$NON-NLS-1$
                 DefaultMessagesImpl.getString("DeleteFolderAction.areYouDeleteFolder")); //$NON-NLS-1$
 
+    }
+    
+    private void delsubFolderForever(IFolder fo) throws CoreException{
+		IResource[] members = fo.members();
+		for (IResource member : members) {
+			if (member instanceof IFolder) {
+				IFolder subFolder = (IFolder) member;
+				if (subFolder.members().length == 0) {
+					LogicalDeleteFileHandle.replaceInFile(
+							LogicalDeleteFileHandle.folderType
+									+ ((IFolder) member).getFullPath().toOSString(), "");
+					subFolder.delete(true, null);
+				} else {
+					delsubFolderForever(subFolder);
+				}
+			}	
+    	}
     }
 
 }
